+++ schemas/src/lib.rs
@@ -10,8 +10,8 @@ pub struct List<T> {
 #[serde(tag = "object", rename = "model")]
 pub struct Model {
     pub id: String,
-    pub running: Option<u64>,
-    pub pending: Option<u64>,
+    pub running: Option<u32>,
+    pub pending: Option<u32>,
     #[serde(flatten)]
     _extra: serde_json::Map<String, serde_json::Value>,
 }
